# internal/handler/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BurnerByte API",
    "description": "Self-hosted temporary email platform API",
    "version": "1.22.0",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "parameters": {
      "orgId": {
        "name": "orgId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "teamId": {
        "name": "teamId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "domainId": {
        "name": "domainId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "page": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 1
        }
      },
      "perPage": {
        "name": "per_page",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 20
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/orgs/{orgId}/domains": {
      "post": {
        "tags": [
          "Domains"
        ],
        "summary": "Add domain",
        "parameters": [
          {
            "$ref": "#/components/parameters/orgId"
          }
        ],
        "responses": {
          "201": {
            "description": "Domain added"
          }
        }
      },
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "List domains",
        "parameters": [
          {
            "$ref": "#/components/parameters/orgId"
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/perPage"
          }
        ],
        "responses": {
          "200": {
            "description": "Domain list"
          }
        }
      }
    },
...
    "/orgs/{orgId}/domains/{domainId}": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Get domain",
        "parameters": [
          {
            "$ref": "#/components/parameters/orgId"
          },
          {
            "$ref": "#/components/parameters/domainId"
          }
        ],
        "responses": {
          "200": {
            "description": "Domain details"
          }
        }
      },
      "patch": {
        "tags": [
          "Domains"
        ],
        "summary": "Update domain",
        "parameters": [
          {
            "$ref": "#/components/parameters/orgId"
          },
          {
            "$ref": "#/components/parameters/domainId"
          }
        ],
        "responses": {
          "200": {
            "description": "Domain updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Domains"
        ],
        "summary": "Delete domain",
        "parameters": [
          {
            "$ref": "#/components/parameters/orgId"
          },
          {
            "$ref": "#/components/parameters/domainId"
          }
        ],
        "responses": {
          "200": {
            "description": "Domain deleted"
          }
        }
      }
    },
...
    "/orgs/{orgId}/teams/{teamId}/domains": {
      "post": {
        "tags": [
          "Domain Assignments"
        ],
        "summary": "Assign domain to team",
        "parameters": [
          {
            "$ref": "#/components/parameters/orgId"
          },
          {
            "$ref": "#/components/parameters/teamId"
          }
        ],
        "responses": {
          "201": {
            "description": "Domain assigned"
          }
        }
      },
      "get": {
        "tags": [
          "Domain Assignments"
        ],
        "summary": "List domain assignments",
        "parameters": [
          {
            "$ref": "#/components/parameters/orgId"
          },
          {
            "$ref": "#/components/parameters/teamId"
          },
          {
            "$ref": "#/components/parameters/page"
          },
          {
            "$ref": "#/components/parameters/perPage"
          }
        ],
        "responses": {
          "200": {
            "description": "Assignment list"
          }
        }
      }
    },
    "/orgs/{orgId}/teams/{teamId}/domains/{domainId}": {
      "patch": {
        "tags": [
          "Domain Assignments"
        ],
        "summary": "Update domain assignment",
        "parameters": [
          {
            "$ref": "#/components/parameters/orgId"
          },
          {
            "$ref": "#/components/parameters/teamId"
          },
          {
            "$ref": "#/components/parameters/domainId"
          }
        ],
        "responses": {
          "200": {
            "description": "Assignment updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Domain Assignments"
        ],
        "summary": "Unassign domain from team",
        "parameters": [
          {
            "$ref": "#/components/parameters/orgId"
          },
          {
            "$ref": "#/components/parameters/teamId"
          },
          {
            "$ref": "#/components/parameters/domainId"
          }
        ],
        "responses": {
          "200": {
            "description": "Domain unassigned"
          }
        }
      }
    },
...
  }
}